Day 11 part 2
One of those days when the most obvious solutions leaves my grasp.

Number of unique numbers is capped and the process is Markov-chainesque
(i.e numbers in next blink are dependent on numbers in previous blink).
So simply maintain a map of numbers in current step and their counts.
For all unique numbers, perform the split/conversion and then add to the
existing counts